added skeleton for event counter tool
git-svn-id: svn+ssh://svn.cern.ch/reps/penn/bjackson/Susy2.0/trunk@22432 4525493e-7705-40b1-a816-d608a930855b

+++ CommonTools/include/EventOutputTool.h
@@ -56,7 +56,7 @@ namespace CommonTools
     ull_t m_event_desc_int;
     ull_t m_sr_helper_int;
 
-    short  m_mc_event_weight;
+    double m_mc_event_weight;
     double m_pileup_weight;
     double m_trigger_weight;
     double m_lepton_weight;
